requirements.txt
vk-api
SQLAlchemy
asyncpg
python-dotenv
psycopg2-binary
aiohttp
httpx